Switch instructions to using GitHub buildpack URL
Since buildkits is deprecated, and the current published version is out
of date (see #61). Whilst we can (and will likely) perform a one-off
publish of `master` to buildkits, it's likely the version there will
become out of date again in the future, so it's just easier to use
GitHub URLs instead.

This buildpack also does not exist on buildpack registry.

Refs #61.